eventbridge-sqs-cdk/cdk/requirements.txt
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-sqs
aws-cdk.core